requirements.txt
Flask==3.0.2
mysql-connector-python==9.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1